Fixed getPerformance command: always request at least one descriptor. Mkisofs 2.02 obsoletes -L, -H, and -P. We now use the long versions. Qt 3.1 compile fix (last time I forgot one cast from ulong long to ulong)
svn path=/branches/k3b_0_11_branch/kdeextragear-1/k3b/; revision=321449

+++ src/device/k3bdiskinfo.cpp
@@ -179,7 +179,7 @@ void K3bCdDevice::NextGenerationDiskInfo::debug() const
 	    << "Tracks:          " << numTracks() << endl
 	    << "Capacity:        " << capacity().toString() 
 	    << " (LBA " << QString::number(capacity().lba())
-	    << ") (" << QString::number(capacity().mode1Bytes()) << " Bytes) (" 
+	    << ") (" << QString::number((unsigned long)capacity().mode1Bytes()) << " Bytes) (" 
 	    << KIO::convertSize(capacity().mode1Bytes()) << ")" << endl
 
 	    << "Remaining size:  " << remainingSize().toString()
